More refactoring and testing
- split away clearsign.c and sock.c
- add unsigned rpm to fixtures
