+++ t/001_sign.t
@@ -3,7 +3,7 @@
 use strict;
 use warnings;
 use bytes;
-use Test::More tests => 12;
+use Test::More tests => 32;
 use File::Temp qw/tempdir/;
 use File::Path qw/remove_tree make_path/;
 use Digest::SHA;
@@ -22,9 +22,7 @@ my $payload  = "test";
 remove_tree($tmp_dir);
 
 make_path($var_dir);
-$ENV{SIGND_VAR} = $var_dir;
 $ENV{LANG} = 'C';
-my $gnupghome=$ENV{GNUPGHOME};
 $ENV{GNUPGHOME} = "$tmp_dir/gnupg";
 make_path($ENV{GNUPGHOME});
 chmod 0700, $ENV{GNUPGHOME};
@@ -46,61 +44,142 @@ my $tmpdir = "$tmp_dir/tmp";
 mkdir($tmpdir, 0700);
 
 my $sign = "./sign --test-sign ./signd";
+my $result;
 
 ###############################################################################
-### if ($cmd eq 'ping') {
-my $ping_result = `$sign -t`;
-is($?, 0, "Checking cmd 'ping' return code");
+### ping
+$result = `$sign -t`;
+is($?, 0, "Checking ping return code");
 
 ###############################################################################
-### if ($cmd eq 'pubkey') {
-my $pubkey_result = `$sign -p`;
+### keyid
+$result = `$sign -k`;
+is($?, 0, "Checking keyid return code");
+
+###############################################################################
+### pubkey
+$result = `$sign -p`;
 is($?, 0, "Checking cmd 'pubkey' return code");
 my $expected = slurp("$fixtures_dir/public-key.asc");
-my @got = split(/\n/, $pubkey_result);
+my @got = split(/\n/, $result);
 splice(@got, 1, 1) if ($got[1] eq 'Version: GnuPG v2');
 my $clean_got = join("\n", @got) . "\n";
 is($clean_got, $expected, "Checking exported pubkey");
 
 ###############################################################################
-### if ($cmd eq 'keygen') {
+### keygen
 my $pubkey = `$sign -P $tmpdir/P -g "rsa\@2048" 800 "$comment" $prj_user`;
 is($?, 0, "Checking cmd 'keygen' return code");
-
 spew("$tmpdir/p", $pubkey);
 system("gpg -q --import $tmpdir/p");
 is($?, 0, "Checking pubkey import return code");
 
 ###############################################################################
-### if ($cmd eq 'certgen') {
+### keyextend
+my $pubkey2 = `$sign -P $tmpdir/P -x 1000 $tmpdir/p`;
+is($?, 0, "Checking cmd 'keyextend' return code");
+spew("$tmpdir/p2", $pubkey2);
+system("gpg -q --import $tmpdir/p2");
+is($?, 0, "Checking pubkey import return code");
+
+###############################################################################
+### certgen
 my $cert = `$sign -P $tmpdir/P -C $tmpdir/p`;
 is($?, 0, "Checking cmd 'certgen' return code");
 my $end_cert = qr/-----END CERTIFICATE-----$/;
 like($cert, $end_cert, "Checking for end of certificate");
 spew("$tmpdir/c", $cert);
-my $openssl_verify = `openssl verify -check_ss_sig -CAfile $tmpdir/c $tmpdir/c`;
+$result = `openssl verify -check_ss_sig -CAfile $tmpdir/c $tmpdir/c`;
 is($?, 0, "Checking openssl verify return code");
+$result = `openssl x509 -in $tmpdir/c -noout -pubkey`;
+is($?, 0, "Checking openssl x509 return code");
+spew("$tmpdir/cp", $result);
 
 ###############################################################################
-### if ($cmd eq 'privsign') {
-spew("$tmpdir/privsign", $payload);
-my $privsign_result = `$sign -P $tmpdir/P -d $tmpdir/privsign`;
-
-is($?, 0, "Checking cmd 'privsign' return code");
-my $verify_privsign = `gpg --verify $tmpdir/privsign.asc 2>&1`;
-like($verify_privsign, qr/Good signature from/, "Checking cmd 'privsign' for 'Good signature'");
-
-###############################################################################
-### if ($cmd eq 'sign') {
+### detached sign
 spew("$tmpdir/sign", $payload);
-my $sign_result = `$sign -d $tmpdir/sign`;
-is($?, 0, "Checking cmd 'sign' return code");
+$result = `$sign -d $tmpdir/sign`;
+is($?, 0, "Checking detached sign return code");
 my $verify_sign = `gpg --verify $tmpdir/sign.asc 2>&1`;
-like($verify_sign, qr/Good signature from/, "Checking for 'Good signature'");
+like($verify_sign, qr/Good signature from/, "Checking detached sign signature");
+unlink("$tmpdir/sign.asc");
+
+###############################################################################
+### detached privsign
+spew("$tmpdir/privsign", $payload);
+$result = `$sign -P $tmpdir/P -d $tmpdir/privsign`;
+is($?, 0, "Checking detached privsign return code");
+$result= `gpg --verify $tmpdir/privsign.asc 2>&1`;
+like($result, qr/Good signature from/, "Checking detached privsign");
+unlink("$tmpdir/privsign.asc");
+
+###############################################################################
+### detached raw sign
+spew("$tmpdir/sign", $payload);
+$result = `$sign -D $tmpdir/sign`;
+is($?, 0, "Checking detached raw sign return code");
+$result = `gpg --verify $tmpdir/sign.sig 2>&1`;
+like($result, qr/Good signature from/, "Checking detached raw sign signature");
+unlink("$tmpdir/sign.sig");
+
+###############################################################################
+### detached privsign
+spew("$tmpdir/privsign", $payload);
+$result = `$sign -P $tmpdir/P -D $tmpdir/privsign`;
+is($?, 0, "Checking detached privsign return code");
+$result = `gpg --verify $tmpdir/privsign.sig 2>&1`;
+like($result, qr/Good signature from/, "Checking detached privsign signature");
+unlink("$tmpdir/privsign.sig");
+
+###############################################################################
+### clear sign
+spew("$tmpdir/sign", $payload);
+$result = `$sign -c $tmpdir/sign`;
+is($?, 0, "Checking clear sign return code");
+$result = `gpg --verify $tmpdir/sign 2>&1`;
+like($result, qr/Good signature from/, "Checking clear sign signature");
+unlink("$tmpdir/sign");
+
+###############################################################################
+### clear privsign
+spew("$tmpdir/privsign", $payload);
+$result = `$sign -P $tmpdir/P -c $tmpdir/privsign`;
+is($?, 0, "Checking clear privsign return code");
+$result = `gpg --verify $tmpdir/privsign 2>&1`;
+like($result, qr/Good signature from/, "Checking clear privsign signature");
+unlink("$tmpdir/privsign");
+
+###############################################################################
+### openssl privsign
+spew("$tmpdir/privsign", $payload);
+$result = `$sign -P $tmpdir/P -h sha256 -O $tmpdir/privsign`;
+is($?, 0, "Checking openssl privsign return code");
+$result = `openssl dgst -verify $tmpdir/cp -signature $tmpdir/privsign.sig $tmpdir/privsign`;
+like($result, qr/Verified OK/, "Checking openssl privsign signature");
+
+###############################################################################
+### rpm sign
+$result = `rpm --dbpath $tmpdir --initdb`;
+$result = `rpm --dbpath $tmpdir --import $fixtures_dir/public-key.asc`;
+spew("$tmpdir/empty.rpm", slurp("$fixtures_dir/empty.rpm"));
+$result = `$sign -h sha256 -r $tmpdir/empty.rpm`;
+is($?, 0, "Checking rpm sign return code");
+$result = `rpm --dbpath $tmpdir --checksig -v $tmpdir/empty.rpm`;
+like($result, qr/^\s*Header V3 RSA\/SHA256 Signature, key ID [0-9a-f]*: OK/m, "Checking rpm header signature");
+like($result, qr/^\s*V3 RSA\/SHA256 Signature, key ID [0-9a-f]*: OK/m, "Checking rpm header+payload signature");
+
+###############################################################################
+### rpm priv sign
+$result = `rpm --dbpath $tmpdir --import $tmpdir/p`;
+spew("$tmpdir/empty.rpm", slurp("$fixtures_dir/empty.rpm"));
+$result = `$sign -P $tmpdir/P -h sha256 -r $tmpdir/empty.rpm`;
+is($?, 0, "Checking rpm privsign return code");
+$result = `rpm --dbpath $tmpdir --checksig -v $tmpdir/empty.rpm`;
+like($result, qr/^\s*Header V3 RSA\/SHA256 Signature, key ID [0-9a-f]*: OK/m, "Checking rpm header signature");
+like($result, qr/^\s*V3 RSA\/SHA256 Signature, key ID [0-9a-f]*: OK/m, "Checking rpm header+payload signature");
 
 ###############################################################################
 ### cleanup
-exit 0;
 remove_tree($tmp_dir);
 exit 0;
